+++ packages/query/src/ast.js
@@ -1,17 +1,15 @@
-
 const defaultOptions = {
-  textFn: v => JSON.stringify(v).toLowerCase()
+  textFn: v => JSON.stringify(v).toLowerCase(),
+  unknownExpressionHandler: expression => {
+    throw new Error(`Unknown expression ${expression.type} at ${expression.col}`)
+  }
 }
 
 const createFilter = (ast, options, cb) => {
-  const mergedOptions = Object.assign({}, defaultOptions, options);
+  const mergedOptions = {...defaultOptions, ...options};
 
-  if (ast.type != 'query') {
-    return cb(new Error('Invalid AST type. Expect type query'));
-  }
-  
   try {
-    const filterFn = queryExpression(ast.value, mergedOptions);
+    const filterFn = queryExpression(ast, mergedOptions);
     cb(null, filterFn);
   } catch (err) {
     cb(err);
@@ -19,44 +17,26 @@ const createFilter = (ast, options, cb) => {
 }
 
 const queryExpression = (exp, options) => {
-  if (exp.type === 'terms') {
-    return termsExpression(exp.value, options);
+  if (exp.type == 'query') {
+    return termsExpression(exp.value, options)
   } else {
-    throw new Error('Invalid AST type. Expect type terms');
+    return termsExpression(exp, options)
   }
 }
 
-const termsExpression = (exps, options) => {
-  const filterFns = exps.map(exp => orExpression(exp, options))
-  
-  let filter = () => true;
-  while (filterFns.length) {
-    const headFns = filterFns.splice(0, Math.min(filterFns.length, 4));
-    
-    let filterFn;
-    if (headFns.length === 4) {
-      filterFn = v => headFns[0](v) && headFns[1](v) && headFns[2](v) && headFns[3](v);
-    } else if (headFns.length === 3) {
-      filterFn = v => headFns[0](v) && headFns[1](v) && headFns[2](v);
-    } else if (headFns.length === 2) {
-      filterFn = v => headFns[0](v) && headFns[1](v);
-    } else {
-      filterFn = headFns[0];
-    }
-    
-    if (filter) {
-      const prevFilter = filter;
-      filter = v => prevFilter(v) && filterFn(v);
-    } else {
-      filter = filterFn;
-    }
+const termsExpression = (exp, options) => {
+  if (exp.type == 'terms') {
+    const filterFns = exp.value.map(e => orExpression(e, options))
+
+    return v => !filterFns.find(filterFn => !filterFn(v))
+  } else {
+    return orExpression(exp, options)
   }
-  return filter;
 }
 
 const orExpression = (exp, options) => {
   if (exp.type === 'or') {
-    const left = andExpression(exp.left, options);
+    const left = orExpression(exp.left, options);
     const right = orExpression(exp.right, options);
     return v => left(v) || right(v)
   } else {
@@ -66,7 +46,7 @@ const orExpression = (exp, options) => {
 
 const andExpression = (exp, options) => {
   if (exp.type === 'and') {
-    const left = notExpression(exp.left, options);
+    const left = andExpression(exp.left, options);
     const right = andExpression(exp.right, options);
     return v => left(v) && right(v);
   } else {
@@ -84,16 +64,22 @@ const notExpression = (exp, options) => {
 }
 
 const expression = (exp, options) => {
-  if (exp.type === 'value') {
+  if (exp.type === 'paren') {
+    return termsExpression(exp.value, options);
+  } else {
+    return valueExpression(exp, options);
+  }
+}
+
+const valueExpression = (exp, options) => {
+  if (exp.type === 'value' || exp.type === 'text') {
     const needle = exp.value.toLowerCase();
     return v => {
       const text = options.textFn(v) || '';
       return text.indexOf(needle) >= 0;
     }
-  } else if (exp.type === 'terms') {
-    return termsExpression(exp.value, options);
   } else {
-    throw new Error(`Unsupported exp type ${exp.type}`);
+    return options.unknownExpressionHandler(exp, options);
   }
 }